pyproject.toml
[tool.poetry]
name = "nowpipes"
version = "0.1.0"
description = ""
authors = ["Maxim Laurijssen <maxim.laurijssen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-box = "^5.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
